fix: trying to use server envs on client
