+++ src/utils/server/builderIO/swc-questionnaire.ts
@@ -1,31 +1,8 @@
 import * as Sentry from '@sentry/nextjs'
-import { nativeEnum, object, string } from 'zod'
 
 import { builderIOClient } from '@/utils/server/builderIO/client'
+import { zodQuestionnaireSchemaValidation } from '@/utils/shared/getSWCQuestionnaire'
 import { getLogger } from '@/utils/shared/logger'
-
-export enum QUESTION_ANSWER_OPTIONS {
-  'Yes' = 'Yes',
-  'No' = 'No',
-  'Not answered' = 'Not answered',
-}
-
-const zodQuestionnaireSchemaValidation = object({
-  data: object({
-    dtsiSlug: string(),
-    q1ExperienceUsingBlockchainTechnology: nativeEnum(QUESTION_ANSWER_OPTIONS),
-    q2BlockchainWillPlayMajorRoleNextInnoWave: nativeEnum(QUESTION_ANSWER_OPTIONS),
-    q3AmerCryptoIsDrivingEconomicGrowth: nativeEnum(QUESTION_ANSWER_OPTIONS),
-    q4UsCompAtRiskIfDigitalAssetsPushedOverse: nativeEnum(QUESTION_ANSWER_OPTIONS),
-    q5UsModernizeRegulatoryEnvironmentForCrypto: nativeEnum(QUESTION_ANSWER_OPTIONS),
-    q6WouldYouVoteInFavorOfLegislation: nativeEnum(QUESTION_ANSWER_OPTIONS),
-    q7VoteInFavorOfLegisToPaymentStablecoins: nativeEnum(QUESTION_ANSWER_OPTIONS),
-    q8ShareAnyOtherOpinionsOnCrypto: string().optional(),
-  }),
-  published: string().regex(/^published$/),
-})
-
-export type SWCQuestionnaireAnswers = Zod.infer<typeof zodQuestionnaireSchemaValidation>
 
 const logger = getLogger(`builderIOQuestionnaire`)
 export async function getQuestionnaire(DTSISlug: string) {
